move circuitopts to transpile folder
